#36 #62 refactor chat into a card for further work
